iw: Fix calculation of fractional multicast rates
Due to missing brackets around the "(int) rate * 10" expression
the float value of "rate" is converted to an integer first and
then multiplied which results for example in "50" instead of the
expected "55" for a give rate of 5.5 Mbps, resulting in an invalid
argument error from iw.

This patch simply puts the multiplication expression in brackets,
so that the type cast is performed on the result and not the
first operand.

Signed-off-by: Jo-Philipp Wich <jow@openwrt.org>